+++ src/app/app.module.ts
@@ -5,6 +5,7 @@ import { NgxWindowModule } from '../../projects/ngx-window/src/public-api';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { LinkComponent } from './link.component';
+import { AdaptivePositioningPageModule } from './modules/adaptive-positioning/adaptive-positioning-page.module';
 import { DropDownPageModule } from './modules/drop-down/drop-down-page.module';
 import { ResizeAndMovePageModule } from './modules/resize-and-move/resize-and-move-page.module';
 
@@ -17,6 +18,7 @@ import { ResizeAndMovePageModule } from './modules/resize-and-move/resize-and-mo
     BrowserModule,
     AppRoutingModule,
     NgxWindowModule,
+    AdaptivePositioningPageModule,
     DropDownPageModule,
     ResizeAndMovePageModule
   ],