+++ src/app/modules/adaptive-positioning/adaptive-positioning-page-routing.module.ts
@@ -1,0 +1,16 @@
+import { NgModule } from '@angular/core';
+import { RouterModule } from '@angular/router';
+import { AdaptivePositioningPageComponent } from './adaptive-positioning-page.component';
+
+@NgModule({
+    imports: [
+        RouterModule.forChild([
+            {
+                path: '',
+                component: AdaptivePositioningPageComponent
+            }
+        ])
+    ],
+    exports: [RouterModule]
+})
+export class AdaptivePositioningPageRoutingModule { }