+++ src/app/modules/adaptive-positioning/adaptive-positioning-page.module.ts
@@ -1,0 +1,17 @@
+import { CommonModule } from '@angular/common';
+import { NgModule } from '@angular/core';
+import { NgxWindowModule } from '../../../../projects/ngx-window/src/public-api';
+import { AdaptivePositioningPageRoutingModule } from './adaptive-positioning-page-routing.module';
+import { AdaptivePositioningPageComponent } from './adaptive-positioning-page.component';
+
+@NgModule({
+    imports: [
+        CommonModule,
+        NgxWindowModule,
+        AdaptivePositioningPageRoutingModule
+    ],
+    declarations: [
+        AdaptivePositioningPageComponent
+    ]
+})
+export class AdaptivePositioningPageModule { }